20191113|like,follow,profile생성

+++ 05_django/04_django_form/articles/views.py
@@ -1,6 +1,7 @@
 import hashlib
 from IPython import embed
 from django.views.decorators.http import require_POST
+from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import Article, Comment
@@ -13,8 +14,8 @@ def index(request):
   # else:
   #   gravatar_url = None
 
-  articles = Article.objects.all()[::-1]
-  context = {'articles':articles}
+  articles = Article.objects.all()
+  context = {'articles':articles,}
   return render(request,'articles/index.html',context)
 
 @login_required
@@ -31,6 +32,8 @@ def create(request):
       # title = form.cleaned_data.get('title')
       # content = form.cleaned_data.get('content')
       # article = Article.objects.create(title=title, content=content)
+      article = form.save(commit = False)
+      article.user = request.user
       article=form.save()
     return redirect('articles:detail', article.pk)
   else:
@@ -46,10 +49,13 @@ def create(request):
 def detail(request,article_pk):
   #article = Article.objects.get(pk=article_pk)
   article = get_object_or_404(Article, pk=article_pk)
+  
+  person = get_object_or_404(get_user_model(), pk=article.user_id)
   comment_form = CommentForm()
   comments = article.comment_set.all()
   context = {
     'article':article,
+    'person':person,
     'comment_form':comment_form,
     'comments':comments,
   }
@@ -57,21 +63,31 @@ def detail(request,article_pk):
 
 @require_POST
 def delete(request, article_pk):
+  # 지금 사용자가 로그인 되어 있는가?
   if request.user.is_authenticated:
+    # 삭제할 게시글 가져옴  
     article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    # 지금 로그인한 사용자와 게시글 작성자 비교
+    if request.user == article.user:
+      article.delete()
+    else:
+      return redirect('articles:detail', article.pk)
   return redirect('articles:index')
   
 @login_required
 def update(request, article_pk):
   article = get_object_or_404(Article, pk=article_pk)
-  if request.method == 'POST':
-    form = ArticleForm(request.POST, instance=article)
-    if form.is_valid():
-      article = form.save()
-      return redirect('articles:detail', article.pk)
-  else :
-    form = ArticleForm(instance=article)
+  if request.user == article.user:
+    if request.method == 'POST':
+      form = ArticleForm(request.POST, instance=article)
+      if form.is_valid():
+        article = form.save()
+        return redirect('articles:detail', article.pk)
+    else :
+      form = ArticleForm(instance=article)
+  else:
+    return redirect('articles:index')
+
   # context로 전달되는 2가지 form 형식
   # 1. GET -> 초기값을 폼에 넣어서 사용자에게 던져줌
   # 2. POST -> is_valid로 False가 리턴됐을 때, 오류 메시지 포함해서 사용자에게 던져줌
@@ -84,15 +100,60 @@ def update(request, article_pk):
 # 댓글 생성 뷰 함수
 @require_POST
 def comment_create(request, article_pk):
-  article = get_object_or_404(Article, pk=article_pk)
+  # article = get_object_or_404(Article, pk=article_pk)
   if request.user.is_authenticated:
     comment_form = CommentForm(request.POST)
     if comment_form.is_valid():
       # save 메서드 -> 선택 인자 : (기본값) commit=True
       # commit=False : DB에 바로 저장되는 것을 막아준다.
       comment = comment_form.save(commit=False)
-      comment.article = article
+      comment.user = request.user
+      comment.article_id = article_pk
       comment.save()
-  return redirect(request.GET.get('next') or 'articles:detail', article.pk)
+  return redirect('articles:detail', article_pk)
 
 
+#댓글 삭제 뷰 함수
+@require_POST
+def comment_delete(request, article_pk, comment_pk):
+  # 1. 로그인 여부 확인
+  if request.user.is_authenticated:
+    comment = get_object_or_404(Comment, pk=comment_pk)
+    # 2. 로그인한 사용자와 댓글 작성자가 같을 경우
+    if request.user == comment.user:
+      comment.delete()
+  return redirect('articles:detail', article_pk)
+
+@login_required
+def like(request, article_pk):
+  # 좋아요 누를 게시글 가져오기
+  article = get_object_or_404(Article, pk=article_pk)
+  # 현재 접속하고 있는 유저
+  user = request.user
+
+  # 현재 게시글을 좋아요 누른 사람 목록에서, 
+  # 현재 접속한 유저가 있을 경우 -> 좋아요 취소
+  if user in article.like_users.all():
+    article.like_users.remove(user)
+  # 목록에 없을 경우 -> 좋아요 누르기
+  else:
+    article.like_users.add(user)
+  return redirect('articles:index')
+
+@login_required
+def follow(request, article_pk, user_pk):
+  # 게시글 작성한 유저
+  person = get_object_or_404(get_user_model(), pk=user_pk)
+  # 지금 접속하고 있는 유저
+  user = request.user
+  # 게시글 작성 유저 팔로워 명단에 접속 중인 유저가 있을 경우
+  # -> 언팔
+  if person != user :
+    if user in person.followers.all():
+      person.followers.remove(user)
+    # 명단에 없으면
+    # -> 팔로우
+    else:
+      person.followers.add(user)
+    # 게시글 상세정보로 redirect
+  return redirect('articles:detail', article_pk)
